+++ pcs_test/tier1/cib_resource/test_clone_unclone.py
@@ -334,23 +334,12 @@ class Clone(
 
     def test_clone_id_is_stonith(self):
         self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_pcs_fail(
+        self.assert_pcs_fail_regardless_of_force(
             "resource clone fence-device".split(),
-            self.stonith_deprecation_warning + fixture_clone_stonith_msg(),
+            "Error: This command does not accept stonith resources.\n",
         )
         self.assert_resources_xml_in_cib(
             fixture_resources_xml(FIXTURE_STONITH_FOR_CLONE)
-        )
-
-    def test_clone_id_is_stonith_forced(self):
-        self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_effect(
-            "resource clone fence-device --force".split(),
-            fixture_resources_xml(FIXTURE_STONITH_CLONE),
-            stderr_full=(
-                self.stonith_deprecation_warning
-                + fixture_clone_stonith_msg(forced=True)
-            ),
         )
 
     def test_clone_group_with_stonith(self):
@@ -484,22 +473,12 @@ class Clone(
 
     def test_promotable_clone_id_is_stonith(self):
         self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_pcs_fail(
+        self.assert_pcs_fail_regardless_of_force(
             "resource promotable fence-device".split(),
-            self.stonith_deprecation_warning + fixture_clone_stonith_msg(),
+            "Error: This command does not accept stonith resources.\n",
         )
         self.assert_resources_xml_in_cib(
             fixture_resources_xml(FIXTURE_STONITH_FOR_CLONE)
-        )
-
-    def test_promotable_clone_id_is_stonith_forced(self):
-        self.set_cib_file(FIXTURE_STONITH_FOR_CLONE)
-        self.assert_pcs_fail(
-            "resource promotable fence-device --force".split(),
-            stderr_start=(
-                self.stonith_deprecation_warning
-                + fixture_clone_stonith_msg(forced=True)
-            ),
         )
 
     def test_promotable_keyword_and_option(self):